% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_boxplot_genedata.R
\name{boxplot_genedata}
\alias{boxplot_genedata}
\title{Fast way to draw boxplot for gene expression}
\usage{
boxplot_genedata(select.genes, design, expr.matrix, convert = T,
  type = c("facet", "normal")[1], contrast = "N.status",
  contrast.list = c("N0", "Np"), genes.levels = NULL, palette = NULL,
  point.alpha = 0.5, box.alpha = 1, method = "wilcox.test",
  map_signif_level = T, x.title = NULL, y.title = NULL,
  legend.position = c(0.9, 0.9), cut = F, ncut = 2, width = 12,
  height = 10, names = "test")
}
\arguments{
\item{select.genes}{the genes you want to plot}

\item{design}{the design object(data of clinical features)}

\item{expr.matrix}{expression matrix of genes}

\item{convert}{Whether to convert Ensembl id as symbols.Default is \code{convert = T}.Note that if you use common names as genes(like miRNAs names),please set \code{convert = F}.}

\item{type}{the style of boxplot.One of "facet" and "normal".}

\item{contrast}{the colnames of contrast value.Like "N.status".}

\item{contrast.list}{the components of contrast.Like c("N0","Np")}

\item{genes.levels}{the level of genes.Default is NULL,and it's recommanded that you set a self-defined levels for better visualization.}

\item{palette}{the color of genes ploted.Default is NULL.}

\item{point.alpha}{the color density of point}

\item{box.alpha}{the color density of box}

\item{method}{the method of P value for comparision.Default is "wilcox.test".See also \code{method} of \code{\link[ggpubr]{compare_means}} function.}

\item{map_signif_level}{boolean value, if the p-value are directly written as annotation or asterisks are used instead. Alternatively one can provide a named numeric vector to create custom mappings from p-values to annotation: For example: c("***"=0.001, "**"=0.01, "*"=0.05)}

\item{x.title}{the names of x axis of plot}

\item{y.title}{the names of y axis of plot}

\item{cut}{whether to cut plot.If the number of genes is very much,it is recommanded to set \code{cut = T} for better visualization.}

\item{ncut}{the number of cut strategy.Only available when \code{cut = T}.Default is 2.}

\item{width}{the width of plot}

\item{height}{the height of plot}

\item{names}{part name of the saved plot}
}
\description{
boxplot_genedata provide fast way to draw boxplot for gene expression based on gene expression matrix and design object
}
\examples{
library(lucky)
data("rna.tpm",package = "lucky")
data("rna.design",package = "lucky")
p <- boxplot_genedata(select.genes = c("ENSG00000004478",
                                       "ENSG00000000457"),
                      design = rna.design,
                      expr.matrix = log2(rna.tpm + 1),
                      contrast = "condition",
                      contrast.list = c("normal","tumor"),
                      genes.levels = NULL,
                      palette=NULL,
                      point.alpha = 0.5,
                      box.alpha = 1,
                      method = "wilcox.test",
                      x.title="Genes",
                      y.title="The Expression of Genes(log TPM)",
                      cut = F,
                      width = 12,height = 10,
                      names = "love")
## See plot of genes expression
print(p$plot)

## See meta data of genes expression
View(p$metadata)
}
\seealso{
\code{\link[ggsignif]{geom_signif}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
