% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{expose}
\alias{expose}
\alias{detach_exposed}
\title{Expose Namespace of a Package}
\usage{
expose(package)

detach_exposed(package)
}
\arguments{
\item{package}{Name of the package.}
}
\description{
Expose all objects (including unexported functions and objects) of the
given package.
}
\examples{
expose('methods')
search()
detach_exposed()
search()
}
