% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_FastCorrelation.R
\name{FastCorrelation}
\alias{FastCorrelation}
\title{Fast way to calculate correlation between variables}
\usage{
FastCorrelation(data, transposition = F, control.markers = NULL,
  target.markers = NULL, method = "pearson")
}
\arguments{
\item{data}{a data frame or a matrix}

\item{control.markers}{control markers}

\item{target.markers}{target markers}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}
}
\value{
a data frame with p value and adjusted p value.
}
\description{
Fast way to calculate correlation between variables
}
\examples{
data <- state.x77;colnames(data)
a <- FastCorrelation(data,
                     transposition = F,
                     control.markers=NULL,
                     target.markers=NULL,
                     method = "pearson")
a1 <- dplyr::filter(a,adj.P.Val < 0.05)

library(ggplot2)
ggplot(as.data.frame(data),aes(x=`Life Exp`,y=`Murder`)) + geom_point() + geom_smooth()
}
\seealso{
\code{\link{cor.test}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
