% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_FastMINE2.R
\name{FastMINE2}
\alias{FastMINE2}
\title{Plus version of MINE correlation}
\usage{
FastMINE2(data, transposition = F, control.markers = NULL,
  target.markers = NULL, alpha = 0.6, nperm = 1000, seed = 1234,
  p.adjust.method = "BH", n.cores = 1)
}
\arguments{
\item{data}{a data frame or a matrix}

\item{control.markers}{control markers}

\item{target.markers}{target markers}

\item{alpha}{In the original article, the authors state that the default value α=0.6 (which is the exponent of the search-grid size B(n)=n^{α}) has been empirically chosen. It is worthwhile noting that alpha and C are defined to obtain an heuristic approximation in a reasonable amount of time. In case of small sample size (etc. n<=63) it is preferable to increase alpha to 1 to obtain a solution closer to the theoretical one. Note: if you set α=1 in a big sample size data set, it would be time-consuming.}

\item{nperm}{integer, number of permutation to perform}

\item{seed}{seed for random number generation reproducibility}

\item{p.adjust.method}{method for pvalue adjustment, see \code{\link{p.adjust}} for available methods.}

\item{n.cores}{the core of parallel calculation}

\item{transpositon}{if you have a data frame like sample on cols and markers on rows(etc. a conventional gene expression matrix), you should set \code{transpositon=T}}
}
\description{
Plus version of MINE correlation.We recommnand \code{FastMINE2} intead of old function \code{FastMINE}.
}
\details{
It seems that even if you set \code{control.markers=NULL} and {target.markers=NULL} at the same time to calculate all pairs, it would not take much time than hours, which is recommanded to do deeper exploration for your data.
}
\examples{
library(minerva)
data("Spellman")
data <- Spellman[1:100] # dim(Spellman)
colnames(data)

dat.cor <- FastMINE2(data)

dat.cor1 <- FastMINE2(data,control.markers = c("time"))

dat.cor2 <- FastMINE2(data,control.markers = c("time","YAL001C"))
}
\seealso{
\code{\link[minerva]{mine}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
