% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Make.cormatrix.R
\name{makeCormatrix}
\alias{makeCormatrix}
\title{conversion between correlation matrix and related long data}
\usage{
makeCormatrix(data, source.col = NULL, target.col = NULL,
  value = NULL, report = T)
}
\arguments{
\item{data}{matrix or data.frame}

\item{source.col}{source colnames}

\item{target.col}{target colnames}

\item{value}{value colnames}

\item{report}{whether to do report}
}
\value{
a long data or a correlation matrix
}
\description{
conversion between correlation matrix and related long data
}
\seealso{
\code{\link[tidyr]{spread}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
