% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_markdownCitation.R
\name{mdCitation}
\alias{mdCitation}
\title{Build markdown citation report}
\usage{
mdCitation(md.path = NULL, csl.path = NULL, bib.path = NULL,
  save.names = NULL, save.type = "docx", sub.bat = F,
  save.path = NULL)
}
\arguments{
\item{md.path}{the path of markdown doc}

\item{csl.path}{the path of .csl file. Default is "american-medical-association.csl" provided by system}

\item{bib.path}{the .bib citation library from cication editor(like Endnote). when you use endnote, you can select "BibTex Export" output form to produce a .txt}

\item{save.names}{the names of saved files}

\item{save.type}{the type of saved files. Default is "docx"}

\item{sub.bat}{whether output indivitual .bat file. Default if \code{FALSE}}

\item{save.path}{the path of output files}
}
\value{
a list
}
\description{
For .md files,when given specified reference form(.csl), citation library(.bib), \code{mardownCitation} product a .bat file to annotate the citations of .md by pandoc.
}
\details{
1. The path should be a absolute path instead of relative one. \cr 2. You have to install pandoc before. More detailed installation please go to \url{https://github.com/jgm/pandoc/releases/tag/2.7.3}
}
\examples{
cit <- mdCitation()
cit <- mdCitation(csl.path = "american-medical-association")
}
\seealso{
csl library \url{https://editor.citationstyles.org/about/}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
