% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_defaults}
\alias{set_defaults}
\title{A (Biased) Collection of Useful Default Options}
\usage{
set_defaults()
}
\description{
The function body is self-explanatory. You may put \code{lucky::set_defaults()}
into your Rprofile.
}
\examples{
print(set_defaults)
}
